Fix constructors

+++ telegrambots-meta/src/main/java/org/telegram/telegrambots/api/methods/send/SendChatAction.java
@@ -49,7 +49,7 @@ public class SendChatAction extends BotApiMethod<Boolean> {
         this.action = checkNotNull(action);
     }
 
-    public SendChatAction(Integer chatId, String action) {
+    public SendChatAction(Long chatId, String action) {
         this.chatId = checkNotNull(chatId).toString();
         this.action = checkNotNull(action);
     }
